Fix a thread sync issue in the event loop
This commit fixes an issue where the event queue in some occations could
throw an exception while pushing an event. The cause of this was the
main loop, which did not wrap calls to the event queue in lock blocks.

The issue was fixed by splitting out parts of the main loop to two
additional methods that checks if the queue is not empty and dequeues
an action respectively.